SDL2_mixer added to the toolchain, finally!
Compiling some of the additional libraries for SDL2_mixer such as libogg/vorbis and/or libmpg123 was a bit of a pain to deal with (e.g. mpg123 refused to compile)

For the time being, I opted to use the MinGW development libraries for SDL2_mixer as opposed to compiling it to make things a bit easier.

I tested SDL2_mixer by compiling a Doom source port (crispy-doom) and it seems to work just fine. Might switch over to compiling SDL2_mixer at some point if I get around to it.